Fixed members links generation code to properly handle :: and ().
The problem became evident after adding documentation that included only :: and () as standalone "words". The code assumed this is also a valid member documentation, even though the name was empty string. The result was that the code replaced all occurences or :: or () with empty reference tag and as a consequence, "real" members which were supposed to be replaced afterwards (and validly started with :: or ended with ()), were left as they were since the original string was no longer found. This is properly handled now, which can be seen from the updated source documentation in the header file as well.