Changed default doxygen handling to automatically treat the first sentence in the documentation as brief description.
Since this option will generate brief description from the first sentence up to the first dot, ending the sentence with three dots actually puts the last two to the detailed section. To allow proper handling, the documentation for testing protocols was changed.

Also added the doxygen configuration file and documentation set info plist to the Xcode project so that we can modify them easily within the Xcode.