Fixed XML makeup handling methods.
Before the original object's NSXmlDocument was passed to each. However this caused problems because some methods actually replace the document in the object's data. To solve it and make it safe for the future additions, the document is no longer passed to the methods but is instead retreived from the passed object data.