Input and output paths options are required.
No real reason for this, except forcing the users to think where the output will be generated - thus making sure nothing is accidentally deleted if --cleanoutput is used. Input will most likely be passed from Xcode $SRCROOT build setting value most often, so it doesn't hurt to force it as well. We'll see.