Implemented global parameters handling.
If Globals.plist file is found in the known templates path, the file values are copied over the factory defaults thus overriding them. This given users a chance to setup their desired global values for all projects. At the moment only reading support is provided, but in the future, the command line may allow users to change the global parameters.

Also refactored the way parameters are stored - instead of using ivar for each, a dictionary is used. This makes it much easier to implement factory defaults and globals. This works similar to NSUserDefaults class.