Implemented --cleantemp switch which removes temporary files.
This option is dynamic, depending the desired utility output. It always leaves the last generated output, but deletes all temporary files used to build it. If the utility creates clean xml, doxygen output is deleted. If the utility creates xhtml, clean xml is deleted as well. If the utility creates and installs documentation set, xhtml and docset in output path are deleted too.

Note that using this option doesn't require --cleanbuild if at least xhtml or documentation set is generated, since all temporary files used to build it will be deleted after succesful build.