All single category sections can be merged into one section in the class.
This option is useful in cases where categories contain several small sections which makes category documentation nicely distributed, but can make the main class very cluttered if the categories documentation are merged into it. Especially if there are many categories for the class. Although this can easily be solved within the documentation itself, it is nicer to allow the user document as they desire and decide the way the documentation is merged at build time. This automatically includes no-pain transition if the documentation grows at a later project stage for example.