Implemented custom object+member references styles.
If a reference is pointing to a member in a different object, the default generated reference (link) description is [member-name object-name]. The user can change this now either either through --objrefstyle command line argument or ObjectReferenceStyle global template option. In both cases the template takes two variables: $MEMBER which will be replaced by the member name and $OBJECT which will be replaced by the object name.