Fixed clean index XML object references handling.
Previously the references were handled by the XSLT which generated the HTML. However this mixed the tasks and required output generation code to be aware of how to create the links. This is properly handled now inside the index XML, so the output generators can simply use the existing information.