Removed all unicode characters inserted by the doxygen custom user scripts.
Although these were invisible in the code unless the file encoding was changed from UTF-8, and were invisible but present in the generated HTML too, it was still better to clean it up, so that any future output generators are not affected by these.

The doxygen custom user scripts were also cleaned up, so they should not produce these chars any more.