Fixed multiline example generation for XHTML.
The problem was that all lines were treated as a single one, wrapping only when reaching the maximum page width. This is properly handled now by wrapping the <div> into a <pre> block. In order to show the same type of font, css had to be tweaked a bit too.