Fixed brief description parsing while generating output.
The problem was with recent refactor which addressed the documentation generation - the code to extract documentation nodes from the parent was moved to a separate method and -extractBriefParagraphsFromItem: and -extractDetailParagraphsFromItem: use it internally. However the brief method "forgot" to return the resulting array of sub nodes... This is fixed now.