Fixed unknown members links handling.
Instead of "blindly" converting all detected member links to the actual references, each possible link is tested. If the given member exists, the reference is generated, otherwise an error is output so that the user can see and correct it. Note that at this moment, verbose output level would probably need to be set to make it easier to pinpoint the error, although project find would also do it anyway.