Refactored output files extension handling.
Instead of hard coding the extension to the clean XML, $EXTENSION placeholder string is used. This is eventually converted to proper extension by concrete GeneratorBase subclasses.

Note that since documentation set generation relies on XHTMLGenerator, it needs to know how to convert the placeholders as well. This brings close coupling between the two objects, but then XHTML generator creates the foundation for the documentation set generation anyway...