Refactored ivars parsing to ignore them alltogether.
This avoids complicating parsing code with complex ivars handling such as blocks for example. Although it would not be too difficult to add this, ivars documentation will most likely not be included in final documentation at all, so it seems it would be overkill to handle parsing. If needed, we can add it later on.