Implemented example before/after requirements testing and handling.
Note that we changed example stringValue from the rest of the paragraph items to use actual string value instead of the value from the source code comment. The reason was to allow simpler debugging as well as exact value in cases where a part of original comment was ignored due to invalid syntax.