Refactored comment objects to allow proper handling for GRMustache.
The most notable change is addition of helper BOOL properties that specify the type of object. As GRMustache doesn't have type comparator, we need to resort to booleans...

Also changed list names for GBCommentParagraph and GBParagraphListItem; instead of using generic "items", we use "paragraphItems" and "listItems" respectively. Although this wouldn't change behavior, it does make the template code more understandable...