GBTask now provides last command line also.
This is quite useful for logging. Note that for any lenghtier process, it might be better to log everything directly to console so that the user can have an indication of progress. This could be enabled if debug logging level is choosen. Will see into this if it would seem viable.