Fixed result handling for GBTask.
The problem was that data received notification was posted when the task finished, even if no data was actually reported. As the code within the notification handler appends the new text after the previous one and emits new line afterwards, this resulted in lastStandardError containing a single \n char. This is properly handled now.