Fixed __attribute__ parsing for property and method definition. Closes #10.
The problem was with __attribute__ being processed as part of property or method definition. This is properly detected now and is ignored, together with any combination of text enclosed within parenthesis following __attribute__ keyword.

For example: "__attribute__()", "__attribute__(something)", "__attribute__((!@#!%^&*~`))" etc. are all ignored, as long as at least one open and close parenthesis should be following __attribute__ keyword