Fixed --help and --version parameters handling.
The problem was with recent addition of templates path handling. An exception is raised if no valid template path is found. However as validation message needs to be sent before the main application handling method due to how DDCli works, exception was raised before we handled any setting. This effectively required the user to pass in a valid templates path even just to see help or version...