Fixed merging method arguments with different var names.
The issue was evident when header and source files used different method argument variable names. We always used the names from the first option parsed. This lead to problems where commend used the other names. For example: "- (void)method:(id)var;" and "- (void)method:(id)theVar;".

To solve the issue, GBMethodData merges argument names too: if only one part is commented, it uses names from that one, otherwise it keeps original names.

Note that this is not 100%; to make it so, we would need to check actual argument names in the comment, but this would complicate the code as we don't have comments parsed at the moment of merging yet. The solution should be good enough though.