Fixed output generation in cases where empty method sections are registered. Closes #26.
Although the section contained no methods and properties, resulting HTML contained the list of all methods registered to the class. The reason was in the way GRMustache handles unknown variables: it descends down the partials stack and tries to use the first object that responds to the given selector (i.e. a variable name). In our case it descends from GBMethodSectionData down to GBMethodsProvider and emmits all it's methods. This situation is properly detected now and all empty sections are removed during processing phase.

Thanks to Derek Clarkson for pointing it out!