Fixed bug that prevented installing documentation set to Xcode. Closes #22.
The problem was in failing installing docset during the process of copying or moving files from output path to Library dir. The reason was as Derek reported in GBOutputGenerator's copyOrMoveItemFromPath:toPath:error: as it tried to remove the destination without checking if it exists. The code now first checks if the destination exists and only deletes it if so.

Have no idea how this escaped me though... Must take some time off or at least stop working on non unit tested code too late in night... :)