Fixed potential duplication of text after cross reference. Closes #44.
The problem was that certain circumstances the text after a cross reference might duplicate the whole paragraph text instead of just getting the remainder. Also added unit test to catch and guard against the bug in the future!

It's strange noone noticed this one as the bug isn't new; in fact the code was there at least since commit 7d6a323e from 2010-11-13... Again thanks to "PrimaryFeather" for reporting it!