Fixed cross reference to members and objects when embedded or followed by punctuation. Closes #33.
appledoc failed to detect cross references if object or member names were surrounded with parenthesis or followed by punctuation. The problem was twofold: to properly handle such cases, matching regex expressions had to be modified a bit plus detection method has to return proper range of detected cross ref so that the remaining text can be properly handled.

Note that similar to #35, the result is all text is properly generated but there are extra spaces. This is going to be covered with #34.