Fixed extra spaces handling before punctuation. Partial fix for #34.
This fixes punctuation and parenthesis, however leaves plural form still prefixed with space. Fixing that will require some redesign of the template. We could use simple solution that would replace single s and es prefixed with space, but that would only work for english and would render some languages incompatible (like s means with in Check and es means is in Spanish). I leave it open for now.