Implemented preparing for docset publishing. Closes #45.
The feature must be enabled by --package-docset and works by using the information from --docset-feed-url which is required by docsetutil when packaging. The result of using it is publish directory in output path with .xar and .atom files. If the .atom file is already there, it is updated if needed, otherwise it is created.