Refactored text and links processing to make it more streamlined.
The algorithm is more straightforward now - the links are only searched at the beginning of the string. If found, link item is created and all text is skipped. If not found the text is marked as text item and the item is generated afterwards. All whitespace is also marked as text item and is added to any skipped text.