Fixed methods parsing in cases where #pragma mark - is used. Closes #50.
The problem was that appledoc considered any - or + as the start of a method. This, coupled with very simple implementation of parsing to keep it as versatile as possible, resulted in this:

	#pragma mark -
	#pragma mark Something

	/** comment */
	- (void)method;

Being recognised as a method `- #pragma mark Something...`. The new parsing code simply checks if the leading - or + is followed by open parenthesis and only recognises the method if so.

Thanks to "tonklon" for bringing this up!