Fixed comment directives processing and normal paragraphs handling.
The problem was when popping directive paragraph which resulted in registering the paragraph to the comment as well if there was no other paragraph on the stack. To properly handle such situations, pushing paragraphs requires sender to specify whether the paragraph is auto-registerable or not. This information is then used when popping and the paragraph is only registered if the flag was set.

This closes another 25 unit tests, with 18 to go.