Fixed remaining text items processing unit tests.
The problem with these was in using normalized description as the string value. Although we could update unit tests to compensate for this, it would be a deviation from the rest of the code - we're now strictly using user's whitespace so we should also use this for string values.

This fixes another set of 4 unit tests with 7 remaining.