Fixed bold-italics decorators handling.
The problem was with the fact that we needed to register text to inner decorator, but register the outer one to the paragraph. To reuse registration code, we correctly assigned inner decorator for text registration, but then only registered the inner decorator to paragraph, effectively discarding outer one completely. This is properly handled now.

In fact, this fixes the remaining 7 unit tests and closes code refactoring!