IMPORTANT: Updated license to modified BSD license!
In short: don't worry, appledoc is still free and will remain so, you can still use it, modify it as well as use generated documentation etc. for any kind of projects - free or commercial!

In long :) I'm primarily developer and don't care about legal stuff, so I didn't take much time to consider this part before. I always wanted to contribute to the community and open the tool to anybody with minimal constraints and will continue to do so in the future. But to prevent potential misuse in today's crazy money-oriented-regardless-of-consequences world :) I took some time to go through at least basic considerations of the way I want to see appledoc go in the future. After all the time I've invested in the tool, this seemed proper to do...

That being said, the main reason for switching from MIT to BSD is clearer text in BSD. But generally both licenses are practically the same and are among most liberal open source licenses :)

I did add one additional requirement to the license and that is to attribute appledoc in generated documentation, your website or similar. I think this is quite sensible and shouldn't interfere with open source-ness of the tool a bit.

If for whatever reason you cannot agree to the license, but want to use appledoc or documentation it generates, feel free to contact me as indicated in the readme file and I'll do my best to help you out and find workable solution!