Fixed Xcode quick help window not displaying documentation. Closes #46.
The problem was in doc set info.plist containing empty keys for unused values. Apparently Xcode requires keys not being present at all in such case. Haven't noticed this in documentation. Thanks to "BloodDragon" for pointing it out.

Note that at this point only quick description and related API is shown, while parameters, declaration and the rest aren't - will need to look into that.

IMPORTANT: template files need to be updated in order for this fix to work!