Fixed "false" warnings for unknown cross references in copied comments as mentioned in previous commit.
Now that was really much simpler than I thought: added boolean flag to GBComment which is set to YES if the comment is copied. Then we only issue warnings if cross reference is not found and the comment is original (i.e. the flag is NO). Note that we only handle @see directives this way; any in-text local cross reference is not going to be reported as it simply won't be matched. And remote references would work the same way in both cases.

This still leaves open the case where both, original and copied objects have members of the same name: in such case copied comment will point to destination members instead of original ones.