Fixed undocumented methods and properties parsing. Closes #57.
The problem was in cases like this:

	/** comment */
	@property (attributes) id commentedProperty;

	- (void)uncommentedMethod;

After working on workaround for #43 (commit 4262bb89029fa3575b4b02ca0de72ed661cf4c5b) forgot to reset comments when parsing methods and properties. Although above case worked properly in case of two methods (in fact there was a unit test specific for that situation and it passed), it didn't work if uncommented method (or property) followed commented property. This is correctly handled now, so it should work for both cases (added another unit test to cover above situation).

Thanks to "BloodDragon" for pointing this one out.