Fixed unneeded template files removal after copying them over to destination. Closes #59.
The problem was when a hidden file was found within a hidden directory (such as .svn with it's contents). In such case the directory was removed before processing the file and when the file was being removed, it no longer existed, so an error was reported. This is properly handled now, thanks to "BloodDragon" for pointing it out.