Fixed short description for comments with single paragraph containing special blocks. Closes #62.
The issue was with comments like this one:

	/** Some text
	
	@warning Description
	*/
	- (void)method;

Although all special blocks (warnings, bugs, examples and lists) are delimited with empty lines from preceding text, they are conceptually part of the same paragraph - they are converted to a paragraph item object and added to the same paragraph as preceding text. The reason for such handling is so that we can add such items to whatever paragraph - like @param or @return description etc. However when creating short description and tooltips for methods, the whole first paragraph (including all special blocks) was used. As from the users point of view, these are clearly delimited with empty line, the resulting documentation contained unexpected tooltips and short descriptions.

These situations are properly handled now and only the text part of the first paragraph up to the first special block is used for short description and tooltip! Thanks to "BloodDragon" for pointing it out!

*NOTE:* after implementing this part I noticed that special blocks were not rendered to the documentation in such cases if `--no-repeat-first-par` option was used. A bit more logic was required in such case, so appledoc now properly repeats omitted parts as a first paragraph of the description paragraphs of a comment!

*IMPORTANT:* This commit requires updating template files!