Fixed detection of known objects regardless of the prefix or suffix.
There's no need to have objects separated with whitespace! Note that this may yield unexpected results in case like this: `[Class(Unknown) method]`. If the Class is found, but no Class(Unknown) category, the Class word will be converted to a cross reference. This should be solvable by requiring prefix though.