Optimized skipped text handling a bit when stepping to next char.
When no link is found at current position, we skip one char and go detect the link there. Instead of appending a single char at the moment, we increment search range and do nothing. Then either when a link is found or end of string is reached, we copy the whole skipped text.