Fixed docsetutil `Entity '' not defined` error.
This error was reported during indexing for any unknown HTML code symbol (for example `&ndash;`). Markdown processor converted plain chars into these symbols as part of the conversion and obviously docsetutil has some issues with it. The workaround was to use plain text for abstract when generating tokens files. This commit is partial fix, the following will also handle some cleaning up of the text to make it more suitable for display.