Fixed (think so :) copied method documentation filename.
When inherited documentation is copied to a method, source information includes both filenames, original and inherited. Previously it sometimes happened that the same name was repeated. This should not happen now anymore.

Note that I have no idea why this was happening - it sometimes did it, sometimes not... Probably due to order of processing since we use NSSet for storing objects to GBStore.