Fixed code blocks auto generated links handling.
If appledoc creates a cross reference link in a text that's later on converted to code block by Markdown processor, all such links are cleaned up when creating HTML value. This is required as Markdown processor doesn't process links inside these blocks.

Note that this is not the most elegant solution: first we process the text and convert to links, then we process generated HTML again and remove inserted links. Ideally, this situation should be detected while processing string and prevent creating links there in the first place. However we would require reimplementing detection of code blocks there, so at least for now, this should be good enough - as code blocks are not that common this should not affect performance too much...