Fixed a bug with `--explicit-crossref` setting being incorrectly assigned from global templates.
The problem was in case global templates included the setting and had it's value set to NO. In such case setExplicitCrossref was sent with proper value, but this method always set matching cross reference markers template to "on", regardless of the value passed. This is properly handled now. Strangely nobody noticed this... Note that command line switches were correctly handled, it was just global templates that were invalid.