Fixed `--repeat-first-par` option handling for methods.
The problem was in GBCommentsProcessor which checked whether current context is static document or top level object and in such case always repeated first paragraph. But when the option was off, the paragraph was repeated for methods as well. The reason was in current context value: in case of method, this is it's parent object, not the method itself! This is properly handled now by introducing another flag for overriding the setting. Not as elegant, but works...

Another option would be to introduce additional argument to `processComment:withContext:store:`, but that would require updating all unit tests, so I ended up with a property.