Implemented project settings. Closes #67.
There are several ways of adding project settings:

- Add AppledocSettings.plist file to one of the paths given to appledoc on command line. If the file is found in a path, it's settings are automatically injected. The same path is also used for source parsing. This allows command line like `appledoc .`.
- Add specific path to a file with arbitrary name and .plist extension. The settings from the file are injected, but the file is not used for parsing.

In both cases, the .plist file has the same syntax as global settings file, except that project settings allow additional option `--input`, which should be an array of strings, each pointing to a source file or directory. These paths can be relative to the .plist file or absolute (will update online documentation soon with examples). The input option is optional; if not used, command line is expected to have at least one source file or directory.