Implemented message file generation when installing documentation set. Closes #71.
The problem was that users expected some files on the `--output` path, however appledoc by default moves files from previous generation phase to next one. So if `--install-docset` is used, output path becomes empty. Although this can be regulated by `--keep-intermediate-files` option, it's confusing as it's off by default.

I chose an user's suggestion to create a text file in output path in such case. To make it even more useful, the file contains the path the documentation set was installed to and installation time.

To make it even less confusing, contents of output path are deleted before running appledoc. Note that although this can be suppressed by an application settings, the setting itself is not wired to a command line switch; having the setting already implemented would make implementing switch later on simpler.