Implemented proper handling of @name sections in more complex circumstances.
Method group sections are now properly detected even when followed by uncommented method. Previously we just assigned last found comment to the method, but now we check if the comment contains method group description and use it to create a new task section. This is more expected.

Note that group matching regex was changed also to match only when found at the start of the comment. This is important as previously and word ending with `name` would be matched as group section...