Fixed manual property getters and setters handling. Closes #75.
The problem was when merging manual getters to property; as selectors were matched, the two methods were merged. However this raised exception as the method types didn't match (it was property vs instance method). This is properly detected now.

Note that this also adds support for custom getters and setters: appledoc now properly matches manual implementation methods to corresponding property as long as standard KVC is used (or `getter=` and `setter=` attributes are used).