Implemented possibility of injecting descriptions to main index file. Closes #74.
Using `--index-desc` switch, you can pass the path to the file containing description that'll be injected in the auto generated index.html. The file can be formatted using the same rules as any other static document. The difference is it's not possible to link to it from other objects or documents (each generated object html already contains the link to main index and hierarchy). But it's possible to link any class or document from within the index description; the same logic is used for that as for any other source code comment or static document.

Online documentation will follow briefly... For now give it a try and let me know if there's anything else missing or something not working as expected.