Fixed partial repeating of comment source strings while processing user's Markdown links. Closes #83.
The problem was in the processing code which didn't use search range but instead searched till the end of the string. This probably came via copy/paste from higher level method which actually handles the whole string... It only affected manual Markdown links written by the user. Also added additional unit test to guard against this for the future changes. Briefly checked other methods dealing with partial strings and they look valid, but it would be nice to keep this in mind in case something similar happens in the future.

Note: This was one of those small issues code-wise that resulted in totally messed up output, giving impression some fundamental flaw was present. But, although a method was fundamentally wrong with it's presumptions, the error was very easy to pinpoint and fix. Now if only all the issues would work like that :)