Fixed remaining bug which resulted in duplicated output. Closes #83.
The error was the same as before: although search range was given when looking for "simple" cross references, string's length was used when computing remaining search range instead of the given source range. This was a bit harder to spot as it only happened in certain circumstances: when using a combination of Markdown pre formatted cross references.

Should work now...