Fixed method merging for cases when instance method is merged on a class method. Closes #90.
This was working when instance method was registered first and class method with the same name was merged on it later on. But it didn't handle the other ordering. Hopefully this will allow handling all cases. Note that there might still be cases when merging properties and class methods - keep in mind if the issue get's re-opened.

Also updated build number to 703.