Added support for UTF8 documentation
Fix appledoc #122 ( https://github.com/tomaz/appledoc/issues/122 )
The issue was that the string sent to Discount was relying on the [str length] method which return the number of characters instead of returning the number of bytes in the UTF8 string. After that, the converted markdown string was converted back to obj-c string using ASCII encoding which was breaking everything.

- Add a test case for UTF8 string handling