Added support for excluding paths from the output.
This is useful if you only want to generate documentation for certain
sources but those sources depend on other sources that should not have
documentation generated for them.  With this feature, you add all
sources as input and exclude certain sources from output generation.  It
will hide warnings on excluded sources, not generate output for them,
but they will be used for merging with your included sources (eg.
inheritance etc.)