Fixed property name detection. Closes #153.
The problem was with recent support for termination macros: originally appledoc always took the last component for property name, but that's obviously not working for termination macros. So the new code took the second component which broke cases like `IBOutlet UIView *name`. To compensate, the code now tries to handle the stuff more intelligently.

**Note:** Just a side note: recent pull requests broke several unit tests, including method/property handling ones, so there might be additional incompatibilities! Will need to check those too!