Fixed html literals handling with doc set. Closes #107.
Today got a feedback from Apple on my bug report:

> When constructing the Tokens.xml file, you have to remember that the file needs to be valid XML first and foremost.  You cannot have a standalone & or an unrecognized XML &entity; in an XML file.  They have to be encoded for the XML first.  For example, if you have the abstract string "This & that", you need to store it in the XML file like this:
>
>     <Abstract>This &amp; that</Abstract>
>
> When docsetutil (or any XML parser) reads the abstract string, the &amp; will get converted into a plain & character.  If you are trying to display a literal entity string, such as "This &asterisk; that", you need to still escape the ampersand character:
>
>     <Abstract>This &amp;asterisk; that</Abstract>
>
> However, if you want your entity to render as a particular character when displayed in QuickHelp, you need to mark the string as HTML:
>
>     <Abstract type="html">This &amp;asterisk; that</Abstract>
>
> The XML parser will see the abstract string as "This &asterisk; that" and will pass it on to the HTML parser, which will then render the string like "This * that".
>
> Please let us know if that information helps you resolve the issue.

Turns out the fix is to change abstract type to html and replace all occurrences of `&` with `&amp;` in token xml files, so for example `&ndash;` should become `&amp;dash;`. Tried few tests on it and works for me. The literals (i.e. `&ndash;`) are properly rendered in Xcode quick help and html. Took a while, but this solution didn't occur to me... Have a go and report back if you have any issue with this.