Fixed category cross references formatting. Closes #160.
The problem was in Discount parsing URL with parenthesis `[text](path/with(parenthesis).html)` as `<a href="path/with(parenthesis">text</a>.html)`. Not sure if this is Markdown rule of how paths with parenthesis are handled, but the simplest solution was to convert category paths to not use parenthesis.

**IMPORTANT:** From here on, category paths are constructed with a plus sign: `class+category`. Previously they were using format `class(category)`. Although this should be simply upgraded by running appledoc, it may break external links. Something to keep in mind...