Fixed NSInvalidArgumentException: -[__NSCFConstantString code]: unrecognized selector sent to instance error. Thanks to @lhunath!
The problem was that appledoc used custom `[NSException raise:format:]` override in an `NSException` category. Overriden method expected `NSError` being passed in as the first parameter, but default method expected `NSString` (i.e. exception name). The fix was to rename override (it's probably always a bad idea to override a method in a category anyway).