Fixed property merging in obscure cases. Closes #184.
If we have this header file:

	@interface X
	@property (readonly, getter=property2) id property1;
	@end

And this .m:

	@implementation X ()
	@property (assign) id property2;
	@end

appledoc would raise exception in `NSParameterAssert` when merging category to main class. To make it work, I commented the assertion out.

Unit tests run successfully in both cases, but I'm leaving the commented code for a while to make sure this doesn't break some other part.