Parsing for objc shortcuts is not less strict. Closes #194.
The problem was in appledoc expecting to have either both or neither of variable name and type keywords for a method argument. So this would pass `-(void)method:`, but this not `-(void)method:var`. The fix is simple: just assume the argument is `id`. So later example would transform internally to `-(void)method:(id)var`. Which is probably what Objective C compiler does too.

Have no idea why I didn't think of this before, it's much less restricting (although probably not the recommended way of doing things in ObjC, but that's whole other topic I guess :)